requirements.txt
# Dev Board Identifier - Requirements
# Install with: pip install -r requirements.txt

# Core ML/CV
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data handling
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# MQTT (optional - for Site Appliance integration)
paho-mqtt>=1.6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Roboflow integration (optional)
roboflow>=1.1.0
